Sửa xong yêu thích và lấy phí ship

+++ BE/petcare-project/src/main/java/com/example/petcareproject/Controller/OrderController.java
@@ -29,7 +29,7 @@ public class OrderController {
         public List<OrderDetailDTO> products;
         public double total;
         public String address;
-        public int shippingCost;
+        public double shippingCost;
         public Long userId; // Thêm userId
         public String paymentMethod;
     }
